Add login and register pages with routing
